Add Serde support to Uris
Adds a Serialize and Deserialize implementation to `Uri`, behind the
`serde` feature flag. There are also a set of tests along side the
implementation, although some work does need to be done to make them
pass as intended

This adds serde_json as a dev-dependency for testing the Serde
implementation.